+++ src/components/card/card.tsx
@@ -1,64 +1,66 @@
+import React, { FunctionComponent } from 'react';
 import styled, { css } from 'styled-components';
 
 export interface CardProps {
-  /**
-   * Padding top
-   */
-  pt?: number;
-  /**
-   * Padding bottom
-   */
-  pb?: number;
-  /**
-   * Padding left
-   */
-  pl?: number;
-  /**
-   * Padding right
-   */
-  pr?: number;
-  /**
-   * Padding horizontal axis
-   */
-  px?: number;
-  /**
-   * Padding vertical axis
-   */
-  py?: number;
-  /**
-   * Should the card elevate on the z-index?
-   */
+  /* Should the card elevate on the z-index? */
   elevate?: boolean;
-  /**
-   * Height of the card
-   */
-  height?: number;
-  /**
-   * Width of the card
-   */
-  width?: number;
+
+  /* Height of the card */
+  height?: string;
+
+  /* Width of the card */
+  width?: string;
+
+  /* Should card have border around it or be entirely flat? */
+  bordered?: boolean;
+
+  /* Content of the card*/
+  children: React.ReactNode;
 }
 
-export const Card = styled.div<CardProps>`
-  height: ${(props) => props.height}px;
-  width: ${(props) => props.width}px;
+const StyledCard = styled.div<CardProps>`
+  height: ${(props) => props.height};
+  width: ${(props) => props.width};
   background-color: #fff;
-  border: 1px solid #dadce0;
+  padding: 1rem;
   border-radius: 8px;
-  padding-top: ${(props) => props.pt}px;
-  padding-bottom: ${(props) => props.pb}px;
-  padding-left: ${(props) => props.pl}px;
-  padding-right: ${(props) => props.pr}px;
-  padding: ${(props) => props.py}px ${(props) => props.px}px;
+  line-height: 1.5;
   ${(props) =>
     props.elevate &&
     css`
       box-shadow: 2px 4px 8px #dadce0;
     `}
+  ${(props) =>
+    props.bordered &&
+    css`
+      border: 1px solid #dadce0;
+    `}
 `;
 
-Card.defaultProps = {
-  height: 400,
-  width: 600,
-  elevate: false,
+export const Card: FunctionComponent<CardProps> = ({
+  children,
+  elevate,
+  height,
+  width,
+  bordered,
+}) => {
+  return (
+    <StyledCard
+      elevate={elevate}
+      height={height}
+      width={width}
+      bordered={bordered}
+    >
+      {children}
+    </StyledCard>
+  );
 };
+
+Card.defaultProps = {
+  width: '400px',
+  height: 'auto',
+  elevate: false,
+  bordered: true,
+};
+
+Card.displayName = 'Card';